Add Splash RecyclerView Item ClickListener
