+++ android/app/src/main/java/kr/pnu/ga2019/presentation/splash/SplashPointAdapter.kt
@@ -10,7 +10,9 @@ import com.bumptech.glide.Glide
 import kr.pnu.ga2019.databinding.LayoutSplashPointBinding
 import kr.pnu.ga2019.domain.entity.Museum
 
-class SplashPointAdapter : RecyclerView.Adapter<SplashPointAdapter.PointViewHolder>() {
+class SplashPointAdapter(
+    private val onClick: ((id: Int) -> Unit) = { /* explicitly empty */ }
+) : RecyclerView.Adapter<SplashPointAdapter.PointViewHolder>() {
 
     private val points: ArrayList<Museum> = arrayListOf()
 
@@ -39,6 +41,7 @@ class SplashPointAdapter : RecyclerView.Adapter<SplashPointAdapter.PointViewHold
     ) : RecyclerView.ViewHolder(binding.root) {
 
         fun bindTo(museum: Museum) {
+            binding.root.setOnClickListener { onClick.invoke(museum.id) }
             binding.pointTitle.text = museum.name
             Glide.with(itemView.context)
                 .load(museum.resId)
